Bitmap util. If else <3
