+++ app/src/main/java/com/example/hrautomation/utils/BitmapUtils/DrawableToBitmapDescriptor.kt
@@ -1,0 +1,20 @@
+package com.example.hrautomation.utils.BitmapUtils
+
+import android.graphics.Bitmap
+import android.graphics.Canvas
+import android.graphics.drawable.Drawable
+import com.google.android.gms.maps.model.BitmapDescriptor
+import com.google.android.gms.maps.model.BitmapDescriptorFactory
+
+class DrawableToBitmapDescriptor {
+    companion object {
+        fun convert(drawable: Drawable): BitmapDescriptor {
+            val canvas = Canvas()
+            val bitmap = Bitmap.createBitmap(drawable.intrinsicWidth, drawable.intrinsicHeight, Bitmap.Config.ARGB_8888)
+            canvas.setBitmap(bitmap)
+            drawable.setBounds(0, 0, drawable.intrinsicWidth, drawable.intrinsicHeight)
+            drawable.draw(canvas)
+            return BitmapDescriptorFactory.fromBitmap(bitmap)
+        }
+    }
+}
